font, viewer menu, & distribution feature added

+++ medousa/sequential/graph/stats/barchart/MyDepthLevelNeighborNodeValueBarChart.java
@@ -62,7 +62,7 @@ public class MyDepthLevelNeighborNodeValueBarChart extends JPanel {
                     }
                 } else {
                     for (String depthNode : MySequentialGraphVars.getSequentialGraphViewer().vc.depthNodeSuccessorMaps.keySet()) {
-                        Map<String, Integer> successorMap = MySequentialGraphVars.getSequentialGraphViewer().vc.depthNodeSuccessorMaps.get(depthNode);
+                        Map<String, Long> successorMap = MySequentialGraphVars.getSequentialGraphViewer().vc.depthNodeSuccessorMaps.get(depthNode);
                         for (String successor : successorMap.keySet()) {
                             String decodedName = (depthNode.contains("x") ? MySequentialGraphSysUtil.getDecodeVariableNodeName(depthNode) : MySequentialGraphSysUtil.decodeNodeName(depthNode));
                             MyNode s = (MyNode) MySequentialGraphVars.g.vRefs.get(successor);
@@ -77,7 +77,7 @@ public class MyDepthLevelNeighborNodeValueBarChart extends JPanel {
                 }
             } else if (MySequentialGraphVars.getSequentialGraphViewer().vc.depthNeighborNodeTypeSelector.getSelectedItem().equals("P.")) {
                 for (String depthNode : MySequentialGraphVars.getSequentialGraphViewer().vc.depthNodePredecessorMaps.keySet()) {
-                    Map<String, Integer> predecessorMap = MySequentialGraphVars.getSequentialGraphViewer().vc.depthNodePredecessorMaps.get(depthNode);
+                    Map<String, Long> predecessorMap = MySequentialGraphVars.getSequentialGraphViewer().vc.depthNodePredecessorMaps.get(depthNode);
                     for (String predecessor : predecessorMap.keySet()) {
                         String decodedName = (depthNode.contains("x") ? MySequentialGraphSysUtil.getDecodeVariableNodeName(depthNode) : MySequentialGraphSysUtil.decodeNodeName(depthNode));
                         this.predecessors.put(decodedName, (float) ((MyNode) MySequentialGraphVars.g.vRefs.get(predecessor)).getNodeDepthInfo(MySequentialGraphVars.currentGraphDepth-1).getOutContribution());
